Modified player(). Added relevant tests.
player() can now accept user inputs to place each ship.
Default values are provided (for testing).
prompt-sync is used so that prompt() can be used in node (for testing).

For simplicity, when testing player('hum'), just press the enter key when asked to enter a value, and the default value will be used.